// src/lib/api/generated/model/referralsQueueListParams.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The canonical backend API for COMPASS clients.
 * OpenAPI spec version: 1.0.0
 */

export type ReferralsQueueListParams = {
/**
 * @minimum 1
 */
page?: number;
/**
 * @minimum 1
 * @maximum 100
 */
page_size?: number;
q?: string;
status?: string;
academic_year?: string;
assignment?: string;
source_type?: string;
reason_category?: string;
order?: string;
};